requirements.txt
click>=8.0.0
prompt-toolkit>=3.0.0
cryptography>=41.0.0
rich>=13.0.0
keyring>=24.0.0
supabase>=2.0.0
pyjwt>=2.8.0
httpx>=0.25.0
python-dotenv>=1.0.0

